draft를 작은 컴포넌트들로 쪼개기
