+++ src/components/pages/post/PostCreationPage.tsx
@@ -5,8 +5,8 @@ import { PostTextEditor } from './PostTextEditor';
 import { PostTitleEditor } from './PostTitleEditor';
 import { PostBackButton } from './PostBackButton';
 import { Loader2, FileText } from 'lucide-react';
-import { DraftStatusIndicator } from './DraftStatusIndicator';
-import { DraftsDrawer } from './DraftsDrawer';
+import { DraftStatusIndicator } from '../draft/DraftStatusIndicator';
+import { DraftsDrawer } from '../draft/DraftsDrawer';
 import { useDraftLoader } from '@/hooks/useDraftLoader';
 import { usePostEditor } from '@/hooks/usePostEditor';
 import { usePostSubmit } from '@/hooks/usePostSubmit';
